+++ src/main/java/com/omar/hex/HexConst.java
@@ -3,15 +3,15 @@ package com.omar.hex;
 import java.awt.*;
 
 public class HexConst {
-    final static Color BGCOLOR =  Color.BLACK;
-    final static Color EMPTYCOLOR =  Color.ORANGE;
-    final static Color GRIDCOLOR =  Color.BLACK;
-    final static Color P1COLOR = new Color( 140, 0, 0);
-    final static Color P1ADJCOLOR = new Color(240,30,30);
-    final static Color P2COLOR = new Color( 0, 140, 0);
-    final static Color P2ADJCOLOR = new Color(30, 240, 30);
-    final static Color TXTCOLOR = Color.BLACK;
-    final static int MAPSIZE = 8;
-    final static int TILESIZE = 60;
-    final static int BORDERS = 15;
+    public final static Color BGCOLOR =  Color.BLACK;
+    public final static Color EMPTYCOLOR =  Color.ORANGE;
+    public final static Color GRIDCOLOR =  Color.BLACK;
+    public final static Color P1COLOR = new Color( 140, 0, 0);
+    public final static Color P1ADJCOLOR = new Color(240,30,30);
+    public final static Color P2COLOR = new Color( 0, 140, 0);
+    public final static Color P2ADJCOLOR = new Color(30, 240, 30);
+    public final static Color TXTCOLOR = Color.BLACK;
+    public final static int MAPSIZE = 8;
+    public final static int TILESIZE = 60;
+    public final static int BORDERS = 15;
 }
