+++ src/main/java/com/omar/model/GameStatus.java
@@ -1,4 +1,4 @@
-package com.omar.hex;
+package com.omar.model;
 
 public enum GameStatus {
     ACTIVE, P1WINS, P2WINS;
