+++ src/main/java/com/omar/model/TurnStatus.java
@@ -1,4 +1,4 @@
-package com.omar.hex;
+package com.omar.model;
 
 public enum TurnStatus {
     P1TURN, P2TURN;
